pyproject.toml
[project]
name = "structural-safety-twin"
version = "0.1.0"
description = "Smart Digital Twin for Structural Health Monitoring - OpenEnv Hackathon"
readme = "README.md"
requires-python = ">=3.14"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "openenv-core>=1.2.0",
    "gymnasium>=1.0.0",
    "numpy>=2.1.0",
    "torch>=2.5.0",
    "sympy>=1.13",
    "typing-extensions>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
    "tensorboard>=2.16.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
# Modern Python linting and formatting
line-length = 88
target-version = "py314"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.setuptools.packages.find]
where = ["."]
include = ["server*", "utils*"]